Delete range command
I created this command with the intent of being better than searching Google for "random number between 1 and 100". While it did "work", the implementation and caveats it had were not worth publishing to users.

The core issue was how it handled large numbers. To support arbitary ranges (e.g. 1 - 100,000,000), generating each number would not be possible. Instead, lazy sequences were used. However, sequences have no indexing, so the implementation fell back to scanning the sequence, which wasn't much better. I tried using `random-sample`, but it doesn't care about the number of results you want and still scans the sequence. So while probability generated the randomness, it was ultimately still slow, and errorous, as it returned no numbers (or below the amount the user requested).

If I can come up with a better implementation, I'll consider adding this command back.