Add poll command and support for components
The poll command is very minimal and could be accomplished manually by
the user, so I'm concerned about the usefulness of this command. I've
chosen to implement it regardless, but it may come into question in
the future. At first, the command used a select menu for "Yes" and
"No", but I decided that using reactions would be a better alternative
since they're simpler, public, and can be amended (with the last point
being debatable as a feature or a bug).