Make interactions pure
This commit makes one simple yet large change to how interactions are processed. Previously, a state machine captured details about an interaction (the actual data and connection) and provided transitions to modify the state (e.g. to send an interaction response). This relied on Tilakone (https://github.com/metosin/tilakone), which was a bit problematic, as the library is minimalistic and merely encapsulated state changes. While having a flow in how to transition to different states (e.g. `:init` -> `:responded` or `:init` -> `:deferred` -> `:responded` / `:realized`), it was still plagued in it having to know "how" to move the state. The solution proposed here makes commands only take interactions (data) and produce "effects" (akin to re-frame effects, but with a vector list instead of a map).

This truly makes interactions unaware in "how" to move states (here, making an HTTP request to Discord and saving the response). If an interaction wants data to continue an operation, they must declare an operation and attach a handler to be called when available (at least, for the current built-in list). The actual processor in "how" is defined one step above in the interaction create event handler

As effects are meant to be simple descriptions, I'd like to keep coupling to certain technology low (e.g. an effect to `:request-user-by-name` rather than an `:http` or `:database-query`). JohnnyJayJay's slash library (https://github.com/JohnnyJayJay/slash) does a similar delegation, except it seems to have no means of responding and continuing (though it's a Ring-style middleware wrap, so it could be easy to program) and relies on macros to handle individual interactions.