Add animanga command
The command currently is very limited and relies on an internal source to fetch the data from. I may replace the source for AniList or MyAnimeList, but the internal source is meant to improve on shortcomings of the previously mentioned services. The real win in this commit is the communication, as Hato (https://github.com/gnarroway/hato) is used (instead of something like clj-http).

Also, I added clj-kondo lints and hooks for Specter and Discljord, so my editor should finally stop complaining about unknown variables (even though we don't use variables in Clojure!).