Add format option to avatar command
The main feature of this commit is the format option on the avatar
command, which allows users to set the image format to return the avatar
in (PNG, JPEG, WebP, and GIF currently). Discord provides a fifth option
(Lottie), but it doesn't seem to work on avatars.

The second feature of this commit is that the avatar command was
refactored to use its own cdn functions instead of the ones Discljord
provides. The namespace (kyleerhabor.assistant.bot.cdn) still relies on
some functions from Discljord, but makes them simpler and more
customizable. The main ways it does that is through:
- returning just the relevant path, as opposed to the entire url (e.g.
  avatars/.../....png instead of https://cdn.discordapp.com/...)
- providing more extension points (e.g. the format) with convenient
  defaults

A big win in this change is that the entire path does not need to be
processed twice by converting it to a URL object, but it still has some
duplication with the parsing of the file name. This is a change I'm not
sure whether or not I want to go beyond and do. I might do it, since it
would allow the functions to become even more general-purpose.

Finally, a small change is that the attach option is internally defined
as `:attach?` instead of just `:attach`, highlighting its conditionality.