Generalize cdn functions
The cdn functions now ask directly for the components they need and
formats it accordingly. This is in contrast to the previous design where
they were convenient by plucking out and processing certain fields to
get the right shape. As a result, the avatar command is more responsible
for processing the components. The main point of this was to derive the
avatar filename as a string without unnecessary computation. This still
involves extra computation, but to a degree that should be insignificant
(just joining a string with a period twice).